add mysql to aws package name
